requirements.txt
pandas
numpy
requests
holidays
fire
coloredlogs
pyyaml